refactor parsing of thumbnail data
