+++ frontend/src/components/main-view/views/login-view/LoginView.tsx
@@ -2,7 +2,6 @@ import React from 'react'
 import {useAppSelector} from '../../../../app/hooks';
 import {selectEmployees} from '../../../../slicer/employeeSlice';
 import {
-    parseEmployeeToLoginThumbnail,
     Views
 } from "../../../../interfaces/IThumbnail";
 
@@ -10,6 +9,7 @@ import {
 import GridView from "../../grid-view/GridView";
 import {selectLoggedIn} from "../../../../slicer/authSlice";
 import {Redirect} from "react-router";
+import {parseEmployeeToLoginThumbnail} from "../../thumbnail/helper";
 
 //interface imports
 
@@ -17,7 +17,7 @@ type Props = {};
 
 function LoginView(props: Props){
     const loggedIn = useAppSelector(selectLoggedIn);
-    const employees = useAppSelector(selectEmployees).map(employee => parseEmployeeToLoginThumbnail(employee));
+    const employees = useAppSelector(selectEmployees).map(parseEmployeeToLoginThumbnail);
     return(
      loggedIn ? <Redirect to={'/start'} /> :  <GridView gridItems={employees} view={Views.LOGIN}/>
     )
