+++ frontend/src/components/main-view/views/products/ProductsView.tsx
@@ -1,13 +1,14 @@
 import React, {useEffect} from 'react'
 import {useAppDispatch, useAppSelector} from '../../../../app/hooks';
 import {getAllProducts, selectProducts} from "../../../../slicer/productSlice";
-import {parseProductToThumbnail, Views} from "../../../../interfaces/IThumbnail";
+import { Views} from "../../../../interfaces/IThumbnail";
 import {selectView} from "../../../../slicer/viewSlice";
 import {productColumns} from "../../list-view/columnDefinition";
 
 //component imports
 import ListView from "../../list-view/ListView";
 import GridView from "../../grid-view/GridView";
+import {parseProductToThumbnail} from "../../thumbnail/helper";
 
 //interface imports
